requirements.txt
asgiref==3.2.9
beautifulsoup4==4.9.1
boto3==1.12.36
botocore==1.15.36
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-htmlmin==0.11.0
django-import-export==2.0.2
django-jazzmin==2.2.7
django-js-asset==1.2.2
django-robots==4.0
django-storages==1.9.1
django-summernote==0.8.11.6
docutils==0.15.2
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.9
jdcal==1.4.1
jmespath==0.9.5
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.0
tablib==1.1.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
